author/requirements.txt
solana
solders
pydantic
pydantic-settings
python-dotenv
